requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
cached-property==1.3.1
click==6.7
Flask==0.12.2
Flask-PW==1.1.3
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
olefile==0.44
packaging==16.8
path.py==10.4
pbr==3.1.1
peewee==2.10.1
peewee-migrate==0.14.0
Pillow==4.2.1
PyMySQL==0.7.11
pyparsing==2.2.0
python-slugify==1.2.4
requests==2.14.2
six==1.10.0
titlecase==0.9.0
Unidecode==0.4.20
Werkzeug==0.12.2
